% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSummaryStatistics.R
\name{calculateSummaryStatistics}
\alias{calculateSummaryStatistics}
\title{Calculate summary statistics}
\usage{
calculateSummaryStatistics(df, value = "value", group = NULL)
}
\arguments{
\item{df}{Data frame to perform calculations on}

\item{value}{Field in the data frame that has the data to calculate.}

\item{group}{Any optional grouping variable}
}
\description{
This function, given a data frame (df) with a continuous variable and a group,
computes distribution statistics.
}
