% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCodeSetTableFromConceptSetExpression.R
\name{createCodeSetTableFromConceptSetExpression}
\alias{createCodeSetTableFromConceptSetExpression}
\title{Resolve Concept Set Expression}
\usage{
createCodeSetTableFromConceptSetExpression(
  connection,
  vocabularyDatabaseSchema,
  conceptSetExpression,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  conceptSetTable = "#concept_sets"
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is
commonly the same as cdmDatabaseSchema. Note that for
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{conceptSetExpression}{A R object (list) that is conforming to OHDSI Circe concept set expression. Usually
from ROhdsiWebApi. Defines the cncept set expression as r list object. This is converted to json and given to
CirceR to get concept set expression sql.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{conceptSetTable}{The name of a remote temp table that has the unique concept_id. Should have column name concept_id and be unique.}
}
\value{
A tibble containing the concept IDs sorted in ascending order.
}
\description{
This function resolves a given concept set expression (R object, not JSON) to
output a temp table with concept IDs.
}
