% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohortFromDrugExposure.R
\name{createCohortFromDrugExposure}
\alias{createCohortFromDrugExposure}
\title{Create Drug exposure cohort}
\usage{
createCohortFromDrugExposure(
  connection = NULL,
  conceptSetExpression,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  drugExposureTable = "#drug_exposure",
  persistenceDay,
  cohortDefinitionId = NULL
)
}
\arguments{
\item{connection}{(optional) An existing database connection object. If NULL, a new connection will be attempted using \code{connectionDetails}.}

\item{conceptSetExpression}{the concept set expression, usually from ROhdsiWebApi}

\item{cdmDatabaseSchema}{The schema name of the CDM database which contains the drug exposure table data.}

\item{tempEmulationSchema}{(optional) The schema used for emulating temporary tables; defaults to the value set in global options.}

\item{drugExposureTable}{drug exposure table}

\item{persistenceDay}{Number of days to check for persistence of drug exposure.}

\item{cohortDefinitionId}{what is the id you would like the final cohort to be?}
}
\description{
This function takes an input a table with person_id, drug_exposure_start_date, drug_exposure_end_date and creates
a cohort table in the form of subject_id, cohort_start_date, cohort_end_date
}
