% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNumeratorCohorts.R
\name{createNumeratorCohorts}
\alias{createNumeratorCohorts}
\title{Create Numerator Cohorts}
\usage{
createNumeratorCohorts(
  connection = NULL,
  conceptSetExpression,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  numeratorCohortTable = "#numerator",
  drugExposureTable = "#drug_exposure",
  persistenceDays = c(0)
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{conceptSetExpression}{the concept set expression, usually from ROhdsiWebApi}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{numeratorCohortTable}{Denominator cohort table.}

\item{drugExposureTable}{The name of the table with "#drug_exposure". Should be a temp table.}

\item{persistenceDays}{Number of days to check for persistence of drug exposure.}
}
\description{
This function takes an input a cohort (called denominator to bind the numerator cohort to), a concept set expression
to create the numerator cohort.
}
