% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateViolinPlots.R
\name{createViolinPlot}
\alias{createViolinPlot}
\title{Create an Enhanced Violin Plot}
\usage{
createViolinPlot(
  data,
  xName = "groupLabel",
  yName = "value",
  fillName = NULL,
  title = "Distribution Plot",
  colorMean = "black",
  colorMedian = "black",
  sizeMean = 3,
  sizeMedian = 3,
  shapeMean = 18,
  shapeMedian = 16,
  angleXLabels = 45
)
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{xName}{The name of the categorical variable (as a string) to plot on the x-axis.}

\item{yName}{The name of the numeric variable (as a string) to plot on the y-axis.}

\item{fillName}{(Optional) The name of the variable (as a string) to use for fill colors. Default is NULL.}

\item{title}{The title of the plot. Default is "Distribution Plot".}

\item{colorMean}{The color to use for the mean symbol. Default is "blue".}

\item{colorMedian}{The color to use for the median symbol. Default is "green".}

\item{sizeMean}{The size of the mean symbol. Default is 3.}

\item{sizeMedian}{The size of the median symbol. Default is 3.}

\item{shapeMean}{The shape of the mean symbol. Default is 18.}

\item{shapeMedian}{The shape of the median symbol. Default is 16.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function generates a violin plot enhanced with overlaid boxplots and
symbols for mean and median values. It uses ggplot2 for plotting, viridis for
color palettes, and hrbrthemes for themes.
}
\examples{
createViolinPlot(data = myData, xName = "group", yName = "value", fillName = "category",
                 title = "My Violin Plot", colorMean = "red", colorMedian = "darkgreen",
                 sizeMean = 4, sizeMedian = 4)

}
