% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCohortGeneratorSubsetOperator.R
\name{defaultCohortGeneratorSubsetOperator}
\alias{defaultCohortGeneratorSubsetOperator}
\title{Default subset operators}
\usage{
defaultCohortGeneratorSubsetOperator(
  minAge = 0,
  maxAge = 90,
  genderSubset = c(8532, 8507),
  restrictToEarliestOccurrence = TRUE,
  race = NULL,
  calendarStartDate = "2016-01-01",
  calendarEndDate = Sys.Date() - 365,
  priorTime = 365,
  followUpTime = 365,
  ethnicity = NULL,
  limitTo = "firstEver"
)
}
\arguments{
\item{minAge}{(optional, set to 0) The min age on first drug exposure. Default is 0.}

\item{maxAge}{(optional, set to 90) The max age on first drug exposure. Default is 90.}

\item{genderSubset}{(Optional) by default set to  c(8532, 8507)}

\item{restrictToEarliestOccurrence}{do you want to restrict to first.}

\item{race}{(Optional) concept id for race}

\item{calendarStartDate}{(optional) The earliest date from which to consider drug exposures. Used to left-censor data.}

\item{calendarEndDate}{(optional) The latest date up to which to consider drug exposures. Used to right-censor data.}

\item{priorTime}{(optional, default 365) The number of days of prior observation.}

\item{followUpTime}{(optional, default 365) The number of days of follow-up observation.}

\item{ethnicity}{(Optional) concept id for ethnicity}

\item{limitTo}{(optional, default 'firstEver') Restrict to first observation period or all observation periods with drug exposure.}
}
\description{
Make a default set of sub-operators
}
