% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposure-package.R, R/DrugExposure.R
\docType{package}
\name{DrugExposure-package}
\alias{DrugExposure-package}
\title{DrugExposure: DrugExposure: To learn patterns of Medication Compliance and Persistence}
\description{
The DrugExposure R package analyzes medication compliance and persistence in cohorts. It assesses drug utilization from first exposure, calculates key metrics like medication possession ratio and proportional days covered, and enhances insights into real-world adherence to improve patient outcomes.

The DrugExposure R package analyzes medication compliance and persistence in cohorts. It assesses drug utilization from first exposure, calculates key metrics like medication possession ratio and proportional days covered, and enhances insights into real-world adherence to improve patient outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DrugExposure/}
  \item \url{https://github.com/OHDSI/DrugExposure}
  \item Report bugs at \url{https://github.com/OHDSI/DrugExposure/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DrugExposure/}
  \item \url{https://github.com/OHDSI/DrugExposure}
  \item Report bugs at \url{https://github.com/OHDSI/DrugExposure/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
