% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEarliestDrugExposureCohort.R
\name{getEarliestDrugExposureCohort}
\alias{getEarliestDrugExposureCohort}
\title{Create Drug exposure cohort}
\usage{
getEarliestDrugExposureCohort(
  connection,
  conceptSetExpression,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputCohortTable = "#de_no_subset",
  outputCohortId = 1,
  conceptSetTable = "#concept_sets",
  drugFirstStartLeftCensorDate = NULL,
  drugFirstStartRightCensorDate = NULL,
  restrictToFirstObservationperiod = TRUE
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{conceptSetExpression}{A R object (list) that is conforming to OHDSI Circe concept set expression. Usually
from ROhdsiWebApi. Defines the cncept set expression as r list object. This is converted to json and given to
CirceR to get concept set expression sql.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{outputCohortTable}{(optional) A temp table that holds the output. This will be created.}

\item{outputCohortId}{(optional) The cohort id of the output cohort. Default 0.}

\item{conceptSetTable}{A temp table that hold the resolved Concept set.}

\item{drugFirstStartLeftCensorDate}{(optional) The earliest date from which to consider drug exposures. Used to left-censor data.}

\item{drugFirstStartRightCensorDate}{(optional) The latest date up to which to consider drug exposures. Used to right-censor data.}

\item{restrictToFirstObservationperiod}{(optional) Default TRUE}
}
\description{
This function takes an input a table with person_id, drug_exposure_start_date, drug_exposure_end_date and creates
a cohort table in the form of subject_id, cohort_start_date, cohort_end_date
}
