% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumeratorCohorts.R
\name{getNumeratorCohorts}
\alias{getNumeratorCohorts}
\title{Create Numerator Cohorts}
\usage{
getNumeratorCohorts(
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  numeratorCohortTableBaseName = "#numerator",
  drugExposureTable = "#drug_exposure",
  gapDays = c(0),
  baseCohortDefinitionId = 100
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{numeratorCohortTableBaseName}{The name of the output table}

\item{drugExposureTable}{The name of the table with "#drug_exposure". Should be a temp table.}

\item{gapDays}{(optional) Number of days to check for persistence of drug exposure. Can take a array of days, and report will be generated for all. This is also called persistence in Atlas, or collapse era days in cohort definition.}

\item{baseCohortDefinitionId}{The minimum cohortId to create cohorts for all gapDays}
}
\description{
This function takes an input a cohort (called denominator to bind the numerator cohort to), a concept set expression
to create the numerator cohort.
}
