% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResolveConceptSetExpression.R
\name{resolveConceptSetExpression}
\alias{resolveConceptSetExpression}
\title{Resolve Concept Set Expression}
\usage{
resolveConceptSetExpression(
  conceptSetExpression,
  connection = NULL,
  connectionDetails = NULL,
  vocabularyDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{conceptSetExpression}{The JSON string of the concept set expression to be resolved.}

\item{connection}{An optional database connection object. If not provided, a connection will be established using the provided connection details.}

\item{connectionDetails}{An optional list specifying the details necessary to establish a database connection. If no connection is provided, this parameter must be specified.}

\item{vocabularyDatabaseSchema}{The name of the database schema containing the vocabulary tables.}

\item{tempEmulationSchema}{The name of the schema used for SQL emulation of temporary tables.}
}
\value{
A tibble containing the concept IDs sorted in ascending order.
}
\description{
This function resolves a given concept set expression (R object, not JSON) to retrieve relevant concept IDs
from a specified vocabulary database schema. It builds and executes a SQL query to extract concept IDs
based on the provided concept set expression.
}
