% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDrugExposure.R
\name{runDrugExposure}
\alias{runDrugExposure}
\title{Run Drug Exposure Analysis}
\usage{
runDrugExposure(
  connectionDetails = NULL,
  connection = NULL,
  conceptSetExpression,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  followUpDays = 365,
  persistenceDays = c(0),
  drugFirstStartLeftCensorDate = NULL,
  drugFirstStartRightCensorDate = NULL
)
}
\arguments{
\item{connectionDetails}{(optional) Database connection details as required by DatabaseConnector package. Used when \code{connection} is NULL.}

\item{connection}{(optional) An existing database connection object. If NULL, a new connection will be attempted using \code{connectionDetails}.}

\item{conceptSetExpression}{Expression that defines the set of drug concepts to analyze. Usually obtained by ROhdsiWebApi from Atlas. This is converted to json and given to CirceR to get concept set expression sql.}

\item{cdmDatabaseSchema}{The schema name of the CDM database which contains the drug exposure table data.}

\item{vocabularyDatabaseSchema}{(optional) The schema name of the vocabulary database. If not provided, it defaults to \code{cdmDatabaseSchema}.}

\item{tempEmulationSchema}{(optional) The schema used for emulating temporary tables; defaults to the value set in global options.}

\item{followUpDays}{Number of days to follow a patient after the initial drug exposure.}

\item{persistenceDays}{(optional) Number of days to check for persistence of drug exposure. Can take a array of days, and report will be generated for all.}

\item{drugFirstStartLeftCensorDate}{(optional) The earliest date from which to consider drug exposures. Used to left-censor data.}

\item{drugFirstStartRightCensorDate}{(optional) The latest date up to which to consider drug exposures. Used to right-censor data.}
}
\description{
This function performs a series of SQL operations to analyze drug exposure data within a specified Common Data Model (CDM) schema. It involves creating temporal tables to capture the initial drug exposure and subsequent follow-up periods based on supplied parameters and predefined SQL logic. The function integrates SQL execution within an R environment by connecting to a database, rendering SQL templates, and executing them.
}
