% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFeatureExtraction.R
\name{runFeatureExtraction}
\alias{runFeatureExtraction}
\title{Run feature extraction}
\usage{
runFeatureExtraction(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortIds,
  cohortTable,
  covariateSettings,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputFolder = NULL,
  aggregated = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where your cohort table exists. cohort table should have columns cohort_definition_id,
subject_id, cohort_start_date, cohort_end_date. The dates for the same subject_id, cohort_definition_id
should be non overlapping. Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortIds}{cohort ids to run feature extraction on. Only one cohort id if aggregated = FALSE.}

\item{cohortTable}{cohort Table Name.}

\item{covariateSettings}{See FeatureExtraction covariateSettings.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{aggregated}{See \code{aggregated} in FeatureExtraction:getDbCovariateData.}
}
\description{
Run feature extraction
}
\details{
This function executes feature extraction
}
