% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCodeSetTableFromConceptSetExpression.R
\name{createCodeSetTableFromConceptSetExpression}
\alias{createCodeSetTableFromConceptSetExpression}
\title{Resolve Concept Set Expression}
\usage{
createCodeSetTableFromConceptSetExpression(
  conceptSetExpression,
  connection,
  vocabularyDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  conceptSetTable = "#concept_sets"
)
}
\arguments{
\item{conceptSetExpression}{The JSON string of the concept set expression to be resolved.}

\item{connection}{An optional database connection object. If not provided, a connection will be established using the provided connection details.}

\item{vocabularyDatabaseSchema}{The name of the database schema containing the vocabulary tables.}

\item{tempEmulationSchema}{The name of the schema used for SQL emulation of temporary tables.}

\item{conceptSetTable}{The table (temp table) with concept_id.}
}
\value{
A tibble containing the concept IDs sorted in ascending order.
}
\description{
This function resolves a given concept set expression (R object, not JSON) to retrieve relevant concept IDs
from a specified vocabulary database schema. It builds and executes a SQL query to extract concept IDs
based on the provided concept set expression.
}
