% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCohortGeneratorSubsetOperator.R
\name{defaultCohortGeneratorSubsetOperator}
\alias{defaultCohortGeneratorSubsetOperator}
\title{Create Default Subset Operators for Cohort Generation}
\usage{
defaultCohortGeneratorSubsetOperator(
  minAge = 0,
  maxAge = 90,
  genderSubset = c(8532, 8507),
  restrictToEarliestOccurrence = TRUE,
  race = NULL,
  calendarStartDate = "2016-01-01",
  calendarEndDate = Sys.Date() - 365,
  priorTime = 365,
  followUpTime = 365,
  ethnicity = NULL,
  limitTo = "firstEver"
)
}
\arguments{
\item{minAge}{Integer, optional. The minimum age at first drug exposure. Defaults to 0 years.}

\item{maxAge}{Integer, optional. The maximum age at first drug exposure. Defaults to 90 years.}

\item{genderSubset}{Numeric vector, optional. Identifies genders using concept IDs. Defaults to c(8532, 8507).}

\item{restrictToEarliestOccurrence}{Logical, optional. Specifies whether to restrict the analysis to the earliest
occurrence of the drug exposure. Defaults to TRUE.}

\item{race}{Integer, optional. Concept ID for specifying race.}

\item{calendarStartDate}{Date, optional. Specifies the earliest date from which to consider drug exposures, used to
left-censor data. Defaults to '2016-01-01'.}

\item{calendarEndDate}{Date, optional. The latest date up to which to consider drug exposures, used to right-censor
data. By default, it is set to one year before the current date to account for claims accrual lag.}

\item{priorTime}{Integer, optional. Represents the number of days of prior observation required before the start of
drug exposure. Defaults to 365 days.}

\item{followUpTime}{Integer, optional. Represents the number of days of follow-up observation after the start of
drug exposure. Defaults to 365 days.}

\item{ethnicity}{Integer, optional. Concept ID for specifying ethnicity.}

\item{limitTo}{Character, optional. Describes whether to restrict the analysis to the first-ever observation period
or to all observation periods with drug exposure. Defaults to 'firstEver'.}
}
\value{
A list of configured subset operators suitable for use in cohort generation functions.
}
\description{
This function configures a set of default subset operators used in cohort generation. It uses CohortGenerator package
subset operator, and allows for easily specifying demographic and temporal parameters to subset a cohort and create
a new cohort.
}
\examples{
defaultCohortGeneratorSubsetOperator(
  minAge = 18,
  maxAge = 65,
  genderSubset = c(8532),
  restrictToEarliestOccurrence = FALSE,
  race = 2106,
  calendarStartDate = as.Date("2010-01-01"),
  calendarEndDate = as.Date("2020-12-31"),
  priorTime = 180,
  followUpTime = 180,
  ethnicity = 38003564,
  limitTo = "all"
)
}
