% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDenominatorCohort.R
\name{getDenominatorCohort}
\alias{getDenominatorCohort}
\title{Create Denominator Cohort for Drug Exposure Analysis}
\usage{
getDenominatorCohort(
  connection,
  conceptSetExpression,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  denominatorCohortTable = "#denominator",
  denominatorCohortId = 1,
  conceptSetTable = "#concept_sets",
  restrictToFirstObservationperiod = TRUE,
  maxFollowUpDays = 365,
  cohortGeneratorSubsetOperators = defaultCohortGeneratorSubsetOperator()
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{conceptSetExpression}{A R object (list) that is conforming to OHDSI Circe concept set expression. Usually
from ROhdsiWebApi. Defines the cncept set expression as r list object. This is converted to json and given to
CirceR to get concept set expression sql.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{denominatorCohortTable}{String, optional. The name of the temporary table to store cohort data. Defaults to "#denominator".}

\item{denominatorCohortId}{Integer, optional. The ID assigned to the denominator cohort. Defaults to 1.}

\item{conceptSetTable}{The name of a remote temp table that has the unique concept_id. Should have column name concept_id and be unique.}

\item{restrictToFirstObservationperiod}{(optional) Default TRUE}

\item{maxFollowUpDays}{(optional, default 365) max number of days to followup the person with continuous observation}

\item{cohortGeneratorSubsetOperators}{(optional) A CohortGenerator Subset operator}
}
\description{
This function constructs a cohort based on the first drug exposure within an observation period. It identifies
the first drug exposure event in either the first (default) or all observation periods for each individual, based on the
concept set expression provided. The start date of this exposure is set as the cohort start date. The cohort end date
is determined by the maximum follow-up days allowed, but it will not exceed the end date of the observation period.
}
