% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDrugExposureInDenominatorCohort.R
\name{getDrugExposureInDenominatorCohort}
\alias{getDrugExposureInDenominatorCohort}
\title{get drug exposure events for a person}
\usage{
getDrugExposureInDenominatorCohort(
  connection = NULL,
  conceptSetExpression,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  conceptSetTable = "#concept_sets",
  denominatorCohortDatabaseSchema = NULL,
  denominatorCohortTable = "#denominator",
  denominatorCohortId = 0,
  drugExposureOutputTable = "#drug_exposure"
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{conceptSetExpression}{A R object (list) that is conforming to OHDSI Circe concept set expression. Usually
from ROhdsiWebApi. Defines the cncept set expression as r list object. This is converted to json and given to
CirceR to get concept set expression sql.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{conceptSetTable}{The name of a remote temp table that has the unique concept_id. Should have column name concept_id and be unique.}

\item{denominatorCohortDatabaseSchema}{(optional) The cohort database schema
that has the denominator cohort.}

\item{denominatorCohortTable}{Denominator cohort table.}

\item{denominatorCohortId}{(optional) The cohort id of the denominator cohort. Default 0.}

\item{drugExposureOutputTable}{the output table}
}
\description{
Given a concept set expression and a denominator cohort to restrict the persons
and exposure event to period in the cohort, this function creates a temp table
that has the records from drug_exposure table.
}
