% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDrugExposure.R
\name{runDrugExposure}
\alias{runDrugExposure}
\title{Run Drug Exposure Analysis}
\usage{
runDrugExposure(
  connectionDetails = NULL,
  connection = NULL,
  conceptSetExpression,
  querySource = TRUE,
  cdmDatabaseSchema,
  denominatorCohortDatabaseSchema,
  denominatorCohortTable,
  denominatorCohortId,
  maxFollowUpDays = 365,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  gapDays = c(0)
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{conceptSetExpression}{A R object (list) that is conforming to OHDSI Circe concept set expression. Usually
from ROhdsiWebApi. Defines the cncept set expression as r list object. This is converted to json and given to
CirceR to get concept set expression sql.}

\item{querySource}{In the drug_exposure table do you want to query drug_source_concept_id. This is set to TRUE by default.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{denominatorCohortDatabaseSchema}{The cohort database schema that has the denominator cohort.}

\item{denominatorCohortTable}{The cohort table name that has the denominator cohort.}

\item{denominatorCohortId}{The cohort id of the denominator cohort.}

\item{maxFollowUpDays}{(optional) Number of days to follow the persons in the denominator cohort.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is
commonly the same as cdmDatabaseSchema. Note that for
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{gapDays}{(optional) Number of days to check for persistence of drug exposure. Can take a array of days, and report will be generated for all. This is also called persistence in Atlas, or collapse era days in cohort definition.}
}
\description{
This function takes as a input a Circe compatible concept set expression (as r list object that
can be converted to json), a denominator cohort or a set of rules to create the denominator
cohort, and checks for occurrence of drug exposure events in the drug_exposure table of
the CDM for the conceptId in the given concept expression in the period and for the subjects in
the denominator cohort. It then computes a series of drug utilization
metrics (adherence, persistence, utilization, patterns)
and reports returns a list of objects that maybe utilized in a drug exposure report.
}
